Reorient boundary between storage and compute
