+++ src/dataflow/src/boundary/boundary.rs
@@ -261,9 +261,9 @@ mod event_link {
     use mz_expr::GlobalId;
     use mz_repr::{Diff, Row};
 
-    use crate::activator::RcActivator;
-    use crate::replay::MzReplay;
-    use crate::server::ActivatedEventPusher;
+    use crate::common::activator::RcActivator;
+    use crate::common::event::ActivatedEventPusher;
+    use crate::common::replay::MzReplay;
 
     use super::{ComputeReplay, StorageCapture};
