+++ src/dataflow/src/boundary/mod.rs
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-pub mod arrangement;
-pub mod compute_state;
-pub mod render;
+mod boundary;
+pub mod tcp_boundary;
+
+pub use boundary::*;
