+++ src/dataflow/src/common/mod.rs
@@ -7,6 +7,6 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-pub mod arrangement;
-pub mod compute_state;
-pub mod render;
+pub(crate) mod activator;
+pub(crate) mod event;
+pub(crate) mod replay;
