+++ src/dataflow/src/logging/reachability.rs
@@ -18,16 +18,16 @@ use timely::dataflow::channels::pact::Exchange;
 use timely::dataflow::operators::capture::EventLink;
 use timely::logging::WorkerIdentifier;
 
-use super::{LogVariant, TimelyLog};
-use crate::activator::RcActivator;
+use mz_dataflow_types::logging::LoggingConfig;
 use mz_dataflow_types::KeysValsHandle;
 use mz_dataflow_types::RowSpine;
-
-use crate::logging::ConsolidateBuffer;
-use crate::replay::MzReplay;
-use mz_dataflow_types::logging::LoggingConfig;
 use mz_ore::iter::IteratorExt;
 use mz_repr::{Datum, Diff, Row, RowArena, Timestamp};
+
+use super::{LogVariant, TimelyLog};
+use crate::common::activator::RcActivator;
+use crate::common::replay::MzReplay;
+use crate::logging::ConsolidateBuffer;
 
 /// Constructs the logging dataflow for reachability logs.
 ///
