+++ src/dataflow/src/logging/timely.rs
@@ -20,15 +20,15 @@ use timely::dataflow::channels::pact::Exchange;
 use timely::dataflow::operators::capture::EventLink;
 use timely::logging::{ParkEvent, TimelyEvent, WorkerIdentifier};
 
-use super::{LogVariant, TimelyLog};
-use crate::activator::RcActivator;
+use mz_dataflow_types::logging::LoggingConfig;
 use mz_dataflow_types::KeysValsHandle;
 use mz_dataflow_types::RowSpine;
-
-use crate::logging::ConsolidateBuffer;
-use crate::replay::MzReplay;
-use mz_dataflow_types::logging::LoggingConfig;
 use mz_repr::{datum_list_size, datum_size, Datum, DatumVec, Diff, Row, Timestamp};
+
+use super::{LogVariant, TimelyLog};
+use crate::common::activator::RcActivator;
+use crate::common::replay::MzReplay;
+use crate::logging::ConsolidateBuffer;
 
 /// Constructs the logging dataflow for timely logs.
 ///
