+++ src/dataflow/src/server.rs
@@ -11,56 +11,37 @@
 
 use std::cell::RefCell;
 use std::collections::HashMap;
-use std::num::NonZeroUsize;
 use std::rc::Rc;
 use std::sync::Mutex;
 use std::time::Instant;
 
 use anyhow::anyhow;
 use crossbeam_channel::TryRecvError;
-use differential_dataflow::trace::cursor::Cursor;
-use differential_dataflow::trace::TraceReader;
 use timely::communication::initialize::WorkerGuards;
 use timely::communication::Allocate;
-use timely::dataflow::operators::unordered_input::UnorderedHandle;
-use timely::dataflow::operators::ActivateCapability;
-use timely::order::PartialOrder;
-use timely::progress::frontier::Antichain;
 use timely::worker::Worker as TimelyWorker;
 use tokio::sync::mpsc;
 
-use mz_dataflow_types::client::Peek;
 use mz_dataflow_types::client::{
     Command, ComputeCommand, ComputeResponse, LocalClient, LocalComputeClient, LocalStorageClient,
     Sender, StorageResponse,
 };
 use mz_dataflow_types::sources::AwsExternalId;
-use mz_dataflow_types::PeekResponse;
 use mz_ore::metrics::MetricsRegistry;
 use mz_ore::now::NowFn;
-use mz_ore::result::ResultExt;
-use mz_repr::{DatumVec, Diff, Row, RowArena, Timestamp};
 
-use crate::compute::arrangement::manager::{TraceBundle, TraceManager, TraceMetrics};
-use crate::event::ActivatedEventPusher;
+use crate::boundary::BoundaryHook;
+use crate::boundary::EventLinkBoundary;
+use crate::boundary::{ComputeReplay, StorageCapture};
+use crate::compute::arrangement::manager::{TraceManager, TraceMetrics};
+use crate::compute::compute_state::ActiveComputeState;
+use crate::compute::compute_state::ComputeState;
 use crate::sink::SinkBaseMetrics;
 use crate::storage::decode::metrics::DecodeMetrics;
 use crate::storage::render::sources::PersistedSourceManager;
 use crate::storage::source::metrics::SourceBaseMetrics;
-
-use crate::server::boundary::BoundaryHook;
-
-pub mod boundary;
-mod compute_state;
-mod storage_state;
-pub mod tcp_boundary;
-
-use crate::server::boundary::EventLinkBoundary;
-use boundary::{ComputeReplay, StorageCapture};
-use compute_state::ActiveComputeState;
-pub(crate) use compute_state::ComputeState;
-use storage_state::ActiveStorageState;
-pub(crate) use storage_state::StorageState;
+use crate::storage::storage_state::ActiveStorageState;
+use crate::storage::storage_state::StorageState;
 
 /// Configures a dataflow server.
 pub struct Config {
@@ -123,7 +104,7 @@ pub fn serve_boundary_requests<
     let (server, storage_client, _compute_client) = serve_boundary(config, create_boundary)?;
     Ok((
         server,
-        crate::server::boundary::BoundaryHook::new(storage_client, requests, workers),
+        crate::boundary::BoundaryHook::new(storage_client, requests, workers),
     ))
 }
 /// Initiates a timely dataflow computation, processing materialized commands.
@@ -406,193 +387,5 @@ where
             Command::Compute(cmd) => self.activate_compute().unwrap().handle_compute_command(cmd),
             Command::Storage(cmd) => self.activate_storage().handle_storage_command(cmd),
         }
-    }
-}
-
-pub struct LocalInput {
-    pub handle: UnorderedHandle<Timestamp, (Row, Timestamp, Diff)>,
-    /// A weak reference to the capability, in case all uses are dropped.
-    pub capability: std::rc::Weak<RefCell<ActivateCapability<Timestamp>>>,
-}
-
-/// An in-progress peek, and data to eventually fulfill it.
-///
-/// Note that `PendingPeek` intentionally does not implement or derive `Clone`,
-/// as each `PendingPeek` is meant to be dropped after it's responded to.
-pub(crate) struct PendingPeek {
-    peek: Peek,
-    /// The data from which the trace derives.
-    trace_bundle: TraceBundle,
-}
-
-impl PendingPeek {
-    /// Produces a corresponding log event.
-    pub fn as_log_event(&self) -> crate::logging::materialized::Peek {
-        crate::logging::materialized::Peek::new(self.peek.id, self.peek.timestamp, self.peek.uuid)
-    }
-
-    /// Attempts to fulfill the peek and reports success.
-    ///
-    /// To produce output at `peek.timestamp`, we must be certain that
-    /// it is no longer changing. A trace guarantees that all future
-    /// changes will be greater than or equal to an element of `upper`.
-    ///
-    /// If an element of `upper` is less or equal to `peek.timestamp`,
-    /// then there can be further updates that would change the output.
-    /// If no element of `upper` is less or equal to `peek.timestamp`,
-    /// then for any time `t` less or equal to `peek.timestamp` it is
-    /// not the case that `upper` is less or equal to that timestamp,
-    /// and so the result cannot further evolve.
-    fn seek_fulfillment(&mut self, upper: &mut Antichain<Timestamp>) -> Option<PeekResponse> {
-        self.trace_bundle.oks_mut().read_upper(upper);
-        if upper.less_equal(&self.peek.timestamp) {
-            return None;
-        }
-        self.trace_bundle.errs_mut().read_upper(upper);
-        if upper.less_equal(&self.peek.timestamp) {
-            return None;
-        }
-        let response = match self.collect_finished_data() {
-            Ok(rows) => PeekResponse::Rows(rows),
-            Err(text) => PeekResponse::Error(text),
-        };
-        Some(response)
-    }
-
-    /// Collects data for a known-complete peek.
-    fn collect_finished_data(&mut self) -> Result<Vec<(Row, NonZeroUsize)>, String> {
-        // Check if there exist any errors and, if so, return whatever one we
-        // find first.
-        let (mut cursor, storage) = self.trace_bundle.errs_mut().cursor();
-        while cursor.key_valid(&storage) {
-            let mut copies = 0;
-            cursor.map_times(&storage, |time, diff| {
-                if time.less_equal(&self.peek.timestamp) {
-                    copies += diff;
-                }
-            });
-            if copies < 0 {
-                return Err(format!(
-                    "Invalid data in source errors, saw retractions ({}) for row that does not exist: {}",
-                    copies * -1,
-                    cursor.key(&storage),
-                ));
-            }
-            if copies > 0 {
-                return Err(cursor.key(&storage).to_string());
-            }
-            cursor.step_key(&storage);
-        }
-
-        // Cursor and bound lifetime for `Row` data in the backing trace.
-        let (mut cursor, storage) = self.trace_bundle.oks_mut().cursor();
-        // Accumulated `Vec<(row, count)>` results that we are likely to return.
-        let mut results = Vec::new();
-
-        // When set, a bound on the number of records we need to return.
-        // The requirements on the records are driven by the finishing's
-        // `order_by` field. Further limiting will happen when the results
-        // are collected, so we don't need to have exactly this many results,
-        // just at least those results that would have been returned.
-        let max_results = self
-            .peek
-            .finishing
-            .limit
-            .map(|l| l + self.peek.finishing.offset);
-
-        if let Some(literal) = &self.peek.key {
-            cursor.seek_key(&storage, literal);
-        }
-
-        let mut row_builder = Row::default();
-        let mut datum_vec = DatumVec::new();
-        let mut l_datum_vec = DatumVec::new();
-        let mut r_datum_vec = DatumVec::new();
-
-        while cursor.key_valid(&storage) {
-            while cursor.val_valid(&storage) {
-                // TODO: This arena could be maintained and reuse for longer
-                // but it wasn't clear at what granularity we should flush
-                // it to ensure we don't accidentally spike our memory use.
-                // This choice is conservative, and not the end of the world
-                // from a performance perspective.
-                let arena = RowArena::new();
-                let key = cursor.key(&storage);
-                let row = cursor.val(&storage);
-                // TODO: We could unpack into a re-used allocation, except
-                // for the arena above (the allocation would not be allowed
-                // to outlive the arena above, from which it might borrow).
-                let mut borrow = datum_vec.borrow_with_many(&[key, row]);
-                if let Some(result) = self
-                    .peek
-                    .map_filter_project
-                    .evaluate_into(&mut borrow, &arena, &mut row_builder)
-                    .map_err_to_string()?
-                {
-                    let mut copies = 0;
-                    cursor.map_times(&storage, |time, diff| {
-                        if time.less_equal(&self.peek.timestamp) {
-                            copies += diff;
-                        }
-                    });
-                    let copies: usize = if copies < 0 {
-                        return Err(format!(
-                            "Invalid data in source, saw retractions ({}) for row that does not exist: {:?}",
-                            copies * -1,
-                            &*borrow,
-                        ));
-                    } else {
-                        copies.try_into().unwrap()
-                    };
-                    // if copies > 0 ... otherwise skip
-                    if let Some(copies) = NonZeroUsize::new(copies) {
-                        results.push((result, copies));
-                    }
-
-                    // If we hold many more than `max_results` records, we can thin down
-                    // `results` using `self.finishing.ordering`.
-                    if let Some(max_results) = max_results {
-                        // We use a threshold twice what we intend, to amortize the work
-                        // across all of the insertions. We could tighten this, but it
-                        // works for the moment.
-                        if results.len() >= 2 * max_results {
-                            if self.peek.finishing.order_by.is_empty() {
-                                results.truncate(max_results);
-                                return Ok(results);
-                            } else {
-                                // We can sort `results` and then truncate to `max_results`.
-                                // This has an effect similar to a priority queue, without
-                                // its interactive dequeueing properties.
-                                // TODO: Had we left these as `Vec<Datum>` we would avoid
-                                // the unpacking; we should consider doing that, although
-                                // it will require a re-pivot of the code to branch on this
-                                // inner test (as we prefer not to maintain `Vec<Datum>`
-                                // in the other case).
-                                results.sort_by(|left, right| {
-                                    let left_datums = l_datum_vec.borrow_with(&left.0);
-                                    let right_datums = r_datum_vec.borrow_with(&right.0);
-                                    mz_expr::compare_columns(
-                                        &self.peek.finishing.order_by,
-                                        &left_datums,
-                                        &right_datums,
-                                        || left.0.cmp(&right.0),
-                                    )
-                                });
-                                results.truncate(max_results);
-                            }
-                        }
-                    }
-                }
-                cursor.step_val(&storage);
-            }
-            // If we had a key, we are now done and can return.
-            if self.peek.key.is_some() {
-                return Ok(results);
-            } else {
-                cursor.step_key(&storage);
-            }
-        }
-
-        Ok(results)
     }
 }
