+++ src/dataflow/src/sink/tail.rs
@@ -47,7 +47,7 @@ where
 
     fn render_continuous_sink(
         &self,
-        compute_state: &mut crate::server::ComputeState,
+        compute_state: &mut crate::compute::compute_state::ComputeState,
         sink: &SinkDesc,
         sink_id: GlobalId,
         sinked_collection: Collection<G, (Option<Row>, Option<Row>), Diff>,
